adds suggested changes and removes types from backend and moves to common
