+++ app/plugins/adp/src/components/ALB/AlbFormFields.ts
@@ -3,13 +3,13 @@ export const albFormFields = [
       label: 'Title',
       name: 'title',
       helperText:
-        'This must be unique - use letters, numbers, or separators such as "_", "-"',
+        'This must be unique',
       validations: {
         required: true,
         pattern: {
-          value: /^[a-zA-Z0-9]+(?:[-_. ][a-zA-Z0-9]+)*$/,
+          value: /^[a-zA-Z0-9]+(?:[. ][a-zA-Z0-9]+)*$/,
           message:
-            'Invalid ALB name format. Use letters, numbers, or "-", "_", "." as separators.',
+            'Invalid ALB name format',
         },
       },
     },
